agent/requirements.txt
livekit-agents>=1.4.0
livekit-plugins-deepgram>=1.4.0
livekit-plugins-groq>=1.4.0
livekit-plugins-cartesia>=1.4.0
livekit-plugins-silero>=1.4.0
